backend/requirements.txt
asgiref==3.5.2
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg==3.1.8
pymongo==3.12.3
pytz==2022.6
sqlparse==0.2.4
typing_extensions==4.5.0
